refactor: move image router and add normalizer
